Add endpoint for deleting job  calls
