+++ lifecycle/lifecycle/endpoints/job.py
@@ -3,13 +3,12 @@ from typing import List
 from fastapi import APIRouter, Request, Response
 from pydantic import BaseModel, Field
 
-from job.async_job_call import save_async_job_call
-from job.async_call import get_async_job_call
-from job.dto_converter import async_job_call_to_dto
 from lifecycle.auth.check import check_auth
 from lifecycle.config import Config
 from lifecycle.config.maintenance import ensure_no_maintenance
 from lifecycle.deployer.redeploy import redeploy_job, reprovision_job, move_job
+from lifecycle.job.async_call import save_async_job_call, get_async_job_call, delete_async_job_call
+from lifecycle.job.dto_converter import async_job_call_to_dto
 from lifecycle.job.ansi import strip_ansi_colors
 from lifecycle.job.graph import build_job_dependencies_graph
 from lifecycle.job.models_registry import update_job_manifest
@@ -167,3 +166,10 @@ def setup_job_endpoints(api: APIRouter, config: Config, plugin_engine: PluginEng
         ensure_no_maintenance()
         check_auth(request, scope=AuthScope.CALL_ADMIN_API)
         save_async_job_call(payload)
+
+    @api.delete('/job/async/call/{call_id}')
+    def _delete_async_job_call(call_id: str, request: Request):
+        """Delete async job call record"""
+        ensure_no_maintenance()
+        check_auth(request, scope=AuthScope.CALL_ADMIN_API)
+        delete_async_job_call(call_id)
