+++ pub/async_test.go
@@ -56,7 +56,7 @@ func TestMultiReplicasAsyncStore(t *testing.T) {
 	}
 	wgPollingRequests.Wait()
 
-	statusCode, responsePayload := getJsonRequest(fmt.Sprintf("http://%v/pub/async/task/%v/poll", addrs[0], "no-such-task"))
+	statusCode, _ := getJsonRequest(fmt.Sprintf("http://%v/pub/async/task/%v/poll", addrs[0], "no-such-task"))
 	assert.Equal(t, statusCode, http.StatusNotFound, "should return status 404 for not existing task")
 
 	for i := 0; i < replicaNum; i++ {
