+++ pub/replicas.go
@@ -13,6 +13,7 @@ import (
 type replicaDiscovery struct {
 	ReplicaDiscoveryHostname string
 	ShouldExit               bool
+	MyAddr                   string
 	otherReplicaAddrs        []string
 	listenPort               string
 }
@@ -46,25 +47,30 @@ func NewStaticReplicaDiscovery(addrs []string) *replicaDiscovery {
 }
 
 func (s *replicaDiscovery) refreshAddrs() error {
-	otherReplicaIPs, err := s.getOtherReplicaIPs()
+	allIps, err := s.getAllReplicaIPs()
+	if err != nil {
+		return errors.Wrap(err, "getting Pub replica IPs")
+	}
+	myIps, err := getMyLocalIPs()
+	if err != nil {
+		return errors.Wrap(err, "reading local IP addresses")
+	}
+	otherReplicaIPs, err := s.getOtherReplicaIPs(allIps, myIps)
 	if err != nil {
 		return err
 	}
 	s.otherReplicaAddrs = MapSlice(otherReplicaIPs, func(ip string) string {
 		return fmt.Sprintf("%s:%v", ip, s.listenPort)
 	})
+	myReplicaIP, err := s.getMyReplicaIP(allIps, myIps)
+	if err != nil {
+		return err
+	}
+	s.MyAddr = fmt.Sprintf("%s:%v", myReplicaIP, s.listenPort)
 	return nil
 }
 
-func (s *replicaDiscovery) getOtherReplicaIPs() ([]string, error) {
-	allIps, err := s.getAllReplicaIPs()
-	if err != nil {
-		return nil, errors.Wrap(err, "getting Pub replica IPs")
-	}
-	myIps, err := getMyLocalIPs()
-	if err != nil {
-		return nil, errors.Wrap(err, "reading local IP addresses")
-	}
+func (s *replicaDiscovery) getOtherReplicaIPs(allIps []string, myIps []string) ([]string, error) {
 	otherIps := []string{}
 	for _, ip := range allIps {
 		if !slices.Contains(myIps, ip) {
@@ -72,6 +78,15 @@ func (s *replicaDiscovery) getOtherReplicaIPs() ([]string, error) {
 		}
 	}
 	return otherIps, nil
+}
+
+func (s *replicaDiscovery) getMyReplicaIP(allIps []string, myIps []string) (string, error) {
+	for _, ip := range allIps {
+		if slices.Contains(myIps, ip) {
+			return ip, nil
+		}
+	}
+	return "", errors.New("Failed to find my replica IP")
 }
 
 func getMyLocalIPs() ([]string, error) {
